refactor: code structure
