+++ src/app_modules/auth/auth_strategies.rs
@@ -1,0 +1,5 @@
+mod base_auth_strategy;
+mod email_password_strategy;
+
+pub use base_auth_strategy::AuthStrategy;
+pub use email_password_strategy::EmailPasswordAuthStrategy;
