+++ src/domain/services.rs
@@ -1,5 +1,7 @@
 mod auth_service;
+mod email_service;
 mod user_service;
 
 pub use auth_service::AuthService;
+pub use email_service::EmailService;
 pub use user_service::UserService;
