+++ src/domain/services/email_service.rs
@@ -1,0 +1,28 @@
+use crate::domain::errors::UserError;
+use tracing::info;
+
+pub struct EmailService;
+
+impl EmailService {
+    pub fn new() -> Self {
+        Self
+    }
+
+    pub fn validate_email(&self, email: &str) -> Result<bool, UserError> {
+        // TODO: Implement email validation logic here
+        Ok(true)
+    }
+
+    pub async fn send_verification_email(
+        &self,
+        email: String,
+        token: String,
+    ) -> Result<(), UserError> {
+        // TODO: Implement email sending logic here
+        info!(
+            "Sending verification email to {} with token {}",
+            email, token
+        );
+        Ok(())
+    }
+}
